go appropriate places when logged out
